+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.3.1] - 2026-09-21
+
+### Fixed
+
+- Fix photos picker URLs #56 @MarcelRobitaille.
+- Fix race condition in photos picker @MarcelRobitaille.
+
 ## [4.3.0] - 2026-09-20
 
 ### Added